Add an allow list to the java implementation
There are valid use cases where someone would want to transliterate a
string, but keep a subset of unicode characters. One example is an
invoice printing system, that supports printing ascii and some unicode
currency characters.

This commit adds overloads to the transliterate methods which allows a
user to supply an array of unicode code points which do not need to be
transliterated.